refine hive related tests
